pyproject.toml
[project]
name = "darkharvest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "ruff>=0.14.10",
]
